requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==2.4.1
attrs==25.4.0
azure-common==1.1.28
azure-core==1.38.2
azure-cosmos==4.14.6
azure-identity==1.25.2
azure-search-documents==11.6.0
azure-storage-blob==12.28.0
beautifulsoup4==4.14.3
blis==1.3.3

cachetools

catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cryptography==46.0.5
cymem==2.0.13
defusedxml==0.7.1
deprecation==2.1.0
devtools==0.12.2
distro==1.9.0
executing==2.2.1
fastapi==0.131.0
fastuuid==0.14.0
filelock==3.24.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
graphrag==3.0.2
graphrag-cache==3.0.2
graphrag-chunking==3.0.2
graphrag-common==3.0.2
graphrag-input==3.0.2
graphrag-llm==3.0.2
graphrag-storage==3.0.2
graphrag-vectors==3.0.2
graspologic-native==1.2.5

gunicorn>=21.2.0

h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
json_repair==0.58.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lancedb==0.24.3
litellm==1.81.14
magika==0.6.3
markdown-it-py==4.0.0
markdownify==1.2.2
markitdown==0.1.5
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.1
murmurhash==1.0.15
nest-asyncio2==1.7.2
networkx==3.6.1
nltk==3.9.1
numpy==2.4.2
onnxruntime==1.24.2
openai==2.21.0
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
preshed==3.0.12
propcache==0.4.1
protobuf==6.33.5
pyarrow==22.0.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pypdfium2==5.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.52.1
sympy==1.14.0
textblob==0.19.0
thinc==8.3.10
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
uvicorn[standard]>=0.27.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websockets==16.0
wrapt==2.1.1
yarl==1.22.0
zipp==3.23.0
